requirements.txt
requests==2.21.0
python-whois==0.7.1
